CHANGE: Major Celestial System Expansion w/ some Ship service, Beyonce service, and Destiny system improvements and fixes along with some expanded debugging features and GM command improvements/new features. Author: Aknor Jaden
---------------------------------
2011-02-24: Patch Initial Release
---------------------------------
This patch basically expands the objects we can spawn in game as well as enables the solar system managers to reload nearly all objects in space that you put there on a server restart. This includes ships (without pilots), drones, most celestial objects with the Celestial and Entity category IDs, Structures (they all appear in their nano-containers), asteroids (they all have same radius because I hard-coded it for now), and all forms of cargo containers: jet-can, secure, audit log, and freight containers.

Mainly, I've been using the /spawn command to test this stuff, but others have used the insider panel and spawned stuff that way. I fixed the /spawn command so that it spawns the right kind of object in the server code according to categoryID and or groupID.

As a side part of this patch, I also did some major work on the Ship "Board", "Eject", "Jettison", and "Launch for corp/self" features. You can now eject from your ship in space, approach another ship and provided you can pilot it, board that ship. You can also board ship to ship without ejecting. You can now jettison items from your cargo into space; a jet-can is spawned and the items are placed inside. You can view the contents of these cans and add/remove items. You can jettison or launch secure/audit log/freight containers into space and view contents/add/remove items. Freight containers do not support removal of items yet as there is a non-existent "aggressionMgr" service that is needed by the client. Another not-so-known kind of container is the Spawn Container. This is what npc and player ship wrecks are and they can be spawned as well, their contents viewed and items added, however, items cannot be removed because they need the same "aggressionMgr" service as the freight containers. There was a problem with setting the name of cans in space such that i had to protect a certain call that was causing the player to "board" the jet can as a ship. This is bypassed for everything except ship types. So, if you rename a can in space, the name will take effect the next time the server restarts.

Here is a list of the celestial objects that have issues as they either do not spawn at all or when they do, there is no graphic element in the game:
* Wormholes, these just show up as icons in space, no graphics (i think there is attribute or some other info missing that needs to be sent to the clients)
* POS Structures, these just appear as their nano-containers; i tried implementing anchor/unanchor effects to no avail. we still don't quite understand what these need
* Deployables, these are the mobile warp disruptors and appear in space as just an icon. if you move to 0m on these, they will "stick" to your ship as you continue to move, i don't know why
* Couds, these are those gas clouds at a site, however the ones called "environments" DO spawn cool looking backgrounds at the location you spawn them

In all this work, I needed to create new source files for some of these new celestial objects, so here is the list of files you'll find in the attached ZIP archive and where you need to put them as they are not included in the patch file itself:
* Structure.h: /include/eve-server/posmgr
* Drone.h: /include/eve-server/ship
* Container.h: /include/eve-server/system
* Deployable.h /include/eve-server/system
* Structure.cpp: /src/eve-server/posmgr
* Drone.cpp: /src/eve-server/ship
* Container.cpp: /src/eve-server/system
* Deployable.cpp: /src/eve-server/system

Other minor cool stuff I fixed/worked on:
* DestinyManager.cpp now has SendCloakShip() and SendUncloakShip() that cloak and uncloak your ship if you call them in the right spots. That was just a test of the effects system and it worked, but i couldn't for the life of me get the ship to cloak after a stargate jump.
* GPoint class now has a MakeRandomPointOnSphere(double radius) function that modifies the existing 3D coordinate to yield a random spot on a sphere of 'radius'
* I added new flags to the packet_types.h file from the recently decompyled Incursion client source code
* Some rather old minor fixes to the chat system, but i don't remember what they were...LOL
* Some attempts to make the Trade between characters work, but it's unfinished as of yet.
* Docking perimeter kinda works as it now checks your ship's distance to station (minus station radius) before allowing you to dock. It still needs to send an error response to the client, but we can do that later.

----------------------------
2011-02-28: Patch Revision 2
----------------------------
*Included in this is a fun little feature of making drones orbit your ship when you spawn them.
*Also, some additional fixes to aid debugging things such as the entire printout of every slash command in the server log, the print out of the Bind Object service whenever a call is made to one of them using a BindID, this way we can start to find out what calls are not implemented in a particular service. The /unspawn and /spawnn slash commands are now fully implemented.
*Finally, there were some bug fixes to prevent server crashing, specifically to the DestinyManager::Orbit() and DestinyManager::Follow() functions as they now properly remove a target when that target is a client and it disconnects from the server.

----------------------------
2011-03-09: Patch Revision 3
----------------------------
* You can now STOP a warp provided your ship has not yet accelerated to 75% of maximum sublight velocity.
* You can now right-click Approach a nearby space object right after undocking from a station or having just jumped through a stargate without having to manually set your speed to non-zero. This is done for you as EVE live does.
* While I have not solved the problem of your ship Stopping upon undock from station the right way, I have "hacked" up something to at least make it do what we want: move away from undock point at full speed. So, until we can figure out how to fix it the right way, at least it does what we want for our convenience.

OTHER NON-DESTINY RELATED FIXES:
* In a brilliant epiphany, I realized that I had programmed POS Structures, Cargo Containers, and Deployables space objects using the wrong Destiny Updates sent to the client, so I fixed this and now if you spawn a POS structure of ANY kind, its fully assembled graphics appear in space, Cargo containers are now "rigid" space objects, and Deployables (mobile warp disruptor bubbles) now actually deploy the warp disrupt bubble graphics. Kinda cool stuff. None of the POS structures "function" of course, but they are pretty to look at. Also, none of the mobile warp disruptor bubbles work to prevent warping of ships inside the bubble, but that will come later. This whole patch was to get objects to appear in space and reappear after a server restart and I've accomplished that. Making these objects function will be something we do later.

----------------------------
2011-03-12: Patch Revision 4
----------------------------
* All Stars, Planets, Moons have a common warp-in point so no matter which direction from which you warp to one of these, you'll end up in roughly the same area with some randomness thrown in so peoples ships don't bump or merge too easily.
* Warping now grabs the Warp Speed Multiplier for your ship type and uses it for warping, so warping "should" be quicker if your ship can warp faster.

+++ src/eve-server/chat/LSCService.cpp
@@ -379,7 +379,11 @@ PyResult LSCService::Handle_JoinChannels(PyCallArgs &call) {
 
 				// Save this subscription to this channel to the database
 				if (!(m_db.IsChannelSubscribedByThisChar(charID, channel->GetChannelID())))
-					m_db.WriteNewChannelSubscriptionToDatabase( charID, channel->GetChannelID() );
+                    m_db.WriteNewChannelSubscriptionToDatabase( charID, channel->GetChannelID(),
+                        call.client->GetCorporationID(), call.client->GetAllianceID(),
+                        2, 0 );     // the "extra" field is hard-coded
+                                                                // to '0' for now since I don't
+                                                                // know what it's used for
 
 				ml->AddItem( chjr.Encode() );
 			}
@@ -436,13 +440,14 @@ PyResult LSCService::Handle_LeaveChannel(PyCallArgs &call) {
     {
         // Remove channel subscription from database if this character was subscribed to it.
         // NOTE: channel subscriptions are NOT saved to the database for private convo chats
-        if (m_db.IsChannelSubscribedByThisChar(call.client->GetCharacterID(),toLeave))
+        if( m_db.IsChannelSubscribedByThisChar(call.client->GetCharacterID(),toLeave) )
             m_db.RemoveChannelSubscriptionFromDatabase(toLeave,call.client->GetCharacterID());
 
         // Remove channel from database if this character was the last one
         // in the channel to leave and it was a private convo (temporary==1):
-        if ((m_channels.find( toLeave )->second->GetMemberCount() == 1)
-            && (m_channels.find( toLeave )->second->GetTemporary() != 0))
+        if( (m_channels.find( toLeave )->second->GetMemberCount() == 1)
+            && (m_channels.find( toLeave )->second->GetTemporary() != 0)
+            && (toLeave >= LSCService::BASE_CHANNEL_ID) )
                 m_db.RemoveChannelFromDatabase(toLeave);
 
         m_channels[ toLeave ]->LeaveChannel( call.client );
@@ -514,13 +519,14 @@ PyResult LSCService::Handle_LeaveChannels(PyCallArgs &call) {
             {
                 // Remove channel subscription from database if this character was subscribed to it.
                 // NOTE: channel subscriptions are NOT saved to the database for private convo chats
-                if (m_db.IsChannelSubscribedByThisChar(call.client->GetCharacterID(),*cur))
-                    m_db.RemoveChannelSubscriptionFromDatabase(*cur,call.client->GetCharacterID());
+                if( m_db.IsChannelSubscribedByThisChar(call.client->GetCharacterID(),*cur))
+                    m_db.RemoveChannelSubscriptionFromDatabase(*cur,call.client->GetCharacterID() );
 
                 // Remove channel from database if this character was the last one
                 // in the channel to leave and it was a private convo (temporary==1):
-                if ((m_channels.find( *cur )->second->GetMemberCount() == 1)
-                    && (m_channels.find( *cur )->second->GetTemporary() != 0))
+                if( (m_channels.find( *cur )->second->GetMemberCount() == 1)
+                    && (m_channels.find( *cur )->second->GetTemporary() != 0)
+                    && (m_channels.find( *cur )->second->GetChannelID() >= LSCService::BASE_CHANNEL_ID) )
                         m_db.RemoveChannelFromDatabase(*cur);
 
                 m_channels[*cur]->LeaveChannel(call.client);
@@ -608,8 +614,12 @@ PyResult LSCService::Handle_CreateChannel( PyCallArgs& call )
 		}
 
 		// Save channel info and channel subscription to the database
-		m_db.WriteNewChannelToDatabase(channel->GetChannelID(),channel->GetDisplayName(),call.client->GetCharacterID(),0);
-		m_db.WriteNewChannelSubscriptionToDatabase(call.client->GetCharacterID(),channel->GetChannelID());
+		m_db.WriteNewChannelToDatabase(channel->GetChannelID(),channel->GetDisplayName(), call.client->GetCharacterID(),0,cmode);
+        m_db.WriteNewChannelSubscriptionToDatabase( call.client->GetCharacterID(), channel->GetChannelID(),
+            call.client->GetCorporationID(), call.client->GetAllianceID(),
+            2, 0 );     // the "extra" field is hard-coded
+                                                    // to '0' for now since I don't
+                                                    // know what it's used for
 
 		channel->JoinChannel( call.client );
 
@@ -695,7 +705,7 @@ PyResult LSCService::Handle_CreateChannel( PyCallArgs& call )
 
 		// Save this channel to the database with the 'temporary' field marked as 1 so that when the last character
 		// leaves this channel, the server knows to remove it from the database:
-		m_db.WriteNewChannelToDatabase(channel_id,call.tuple->GetItem(0)->AsString()->content(),call.client->GetCharacterID(),1);
+		m_db.WriteNewChannelToDatabase(channel_id,call.tuple->GetItem(0)->AsString()->content(),call.client->GetCharacterID(),1,cmode);
 	}
 
 
@@ -706,7 +716,11 @@ PyResult LSCService::Handle_CreateChannel( PyCallArgs& call )
 		{
 			// Save this subscription to this channel to the database IF it is not temporary:
 			if (channel->GetTemporary() == 0)
-				m_db.WriteNewChannelSubscriptionToDatabase( call.client->GetCharacterID(), channel->GetChannelID() );
+                m_db.WriteNewChannelSubscriptionToDatabase( call.client->GetCharacterID(), channel->GetChannelID(),
+                    call.client->GetCorporationID(), call.client->GetAllianceID(),
+                    2, 0 );     // the "extra" field is hard-coded
+                                                            // to '0' for now since I don't
+                                                            // know what it's used for
 
 			channel->JoinChannel( call.client );
 
@@ -917,6 +931,10 @@ PyResult LSCService::Handle_DestroyChannel( PyCallArgs& call )
     // packet sent back to the client?
     // **************************
 
+    // Remove the channel from the database:
+    m_db.RemoveChannelFromDatabase( res->second->GetChannelID() );
+
+    // Finally, remove the channel from the server dynamic objects:
     res->second->Evacuate( call.client );
     SafeDelete( res->second );
     m_channels.erase( res );
@@ -993,6 +1011,13 @@ PyResult LSCService::Handle_AccessControl( PyCallArgs& call )
 	// -- Aknor Jaden (2010-11-26)
 
 	int32 channel_id = 0;
+
+    // BIG TODO:  The whole reason why normal players cannot post chats in other channels has to do with the Access Mode
+    // in the channel settings dialog in the client.  Now, I don't know why chatting in the Help/Rookie Help is not allowed
+    // for ANY normal players, however, at the very least, implementing the change in Access Mode to 3 = Allowed may fix
+    // the issue.  The only thing to figure out is how to format a packet to be sent to the client(s) to inform of the change
+    // in access mode.  Is this really needed though, since the owner will change the mode and the owner's client will know
+    // immediately, and anyone wanting to join will get that mode value when the JoinChannel has been called.
 
     // call.tuple->GetItem(0)->AsInt()->value() = channel ID
     // call.tuple->GetItem(1)->IsNone() == true  <---- change made to "" field
@@ -1097,10 +1122,6 @@ PyResult LSCService::Handle_Invite(PyCallArgs &call)
             //chatInvitePacket.integer5 = 1;
             //PyTuple *tuple = chatInvitePacket.Encode();
             //entityList().Unicast(invited_char_ID, "", "", &tuple, false);
-
-			// Save this subscription to this channel to the database IF it is not temporary:
-			if ((channel->GetTemporary() == 0) && (m_db.IsChannelSubscribedByThisChar(invited_char_ID, channel_ID) == false))
-				m_db.WriteNewChannelSubscriptionToDatabase( invited_char_ID, channel_ID );
 		}
 		else
 		{
